Fixed Eclipse plugin install issues

+++ Plugins/Aspose-Cells-Java-for-Eclipse(Maven)/AsposeCellsEclipsePlugin/src/com/aspose/cells/maven/examples/AsposeExampleSupport.java
@@ -74,9 +74,15 @@ public class AsposeExampleSupport implements IRunnableWithProgress {
 				+ File.separator + AsposeConstants.MAVEN_POM_XML;
 
 		try {
-			FileUtils.copyDirectory(srcExampleCategoryPath, destExampleCategoryPath);
-			Files.copy(srcUtil, destUtil, StandardCopyOption.REPLACE_EXISTING);
-			FileUtils.copyDirectory(srcExampleResourceCategoryPath, destExampleResourceCategoryPath);
+			if(srcExampleCategoryPath.exists()) {
+				FileUtils.copyDirectory(srcExampleCategoryPath, destExampleCategoryPath);	
+			}					
+			if(srcUtil.toFile().exists()) {
+				Files.copy(srcUtil, destUtil, StandardCopyOption.REPLACE_EXISTING);	
+			}			
+			if(srcExampleResourceCategoryPath.exists()) {
+				FileUtils.copyDirectory(srcExampleResourceCategoryPath, destExampleResourceCategoryPath);				
+			}
 
 			NodeList examplesNoneAsposeDependencies = AsposeMavenProjectManager.getInstance()
 					.getDependenciesFromPOM(repositoryPOM_XML, AsposeConstants.ASPOSE_GROUP_ID);
